fix:fixed table ui and life list title
